agent/requirements.txt
langchain==0.3.13
langchain-google-genai==2.0.8
langchain-openai==0.2.14
langgraph==0.2.62
langserve==0.3.0
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
httpx==0.27.2
sse-starlette==3.0.3
starlette==0.41.3
